[graphite-web] Define a plan for the graphite-web service
At least in the first draft, the Python module installations are baked into
this plan, but we should probably split them out into their own plans.

In some cases, e.g. uwsgi, there are plans in the Habitat python2 org, but none
of the plans there are being tagged for the stable channel. If they ever start,
we can look at converting.