[inspec] Use chef/inspec as the build source
This gets us version upgrades for free instead of having to wire up an IFTTT
rule for Inspec. It also saves time in the builds since we're just copying
files instead of rebuilding the gem.

It also gets us to a version with fixes for the earlier issues with Busybox's
`ps` and `netstat` commands.